[MISC] move type_traits/lazy to utility
