+++ include/seqan3/alignment/matrix/detail/coordinate_matrix.hpp
@@ -21,8 +21,8 @@
 #include <seqan3/core/simd/simd_algorithm.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
 #include <seqan3/core/simd/view_iota_simd.hpp>
-#include <seqan3/core/type_traits/lazy.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
+#include <seqan3/utility/type_traits/lazy_conditional.hpp>
 
 namespace seqan3::detail
 {
