+++ include/seqan3/alignment/pairwise/align_result_selector.hpp
@@ -26,11 +26,11 @@
 #include <seqan3/core/algorithm/configuration.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/core/range/type_traits.hpp>
-#include <seqan3/core/type_traits/lazy.hpp>
 #include <seqan3/range/decorator/gap_decorator.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
 #include <seqan3/utility/type_traits/basic.hpp>
 #include <seqan3/utility/type_traits/detail/transformation_trait_or.hpp>
+#include <seqan3/utility/type_traits/lazy_conditional.hpp>
 
 namespace seqan3::detail
 {
