+++ include/seqan3/alignment/pairwise/alignment_configurator.hpp
@@ -56,9 +56,9 @@
 #include <seqan3/core/detail/deferred_crtp_base.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/core/simd/simd.hpp>
-#include <seqan3/core/type_traits/lazy.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
 #include <seqan3/range/views/zip.hpp>
+#include <seqan3/utility/type_traits/lazy_conditional.hpp>
 
 namespace seqan3::detail
 {
