+++ include/seqan3/alignment/pairwise/detail/policy_optimum_tracker_simd.hpp
@@ -19,8 +19,8 @@
 #include <seqan3/core/simd/simd.hpp>
 #include <seqan3/core/simd/simd_algorithm.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
-#include <seqan3/core/type_traits/lazy.hpp>
 #include <seqan3/range/views/zip.hpp>
+#include <seqan3/utility/type_traits/lazy_conditional.hpp>
 
 namespace seqan3::detail
 {
