+++ include/seqan3/alignment/pairwise/detail/type_traits.hpp
@@ -34,10 +34,10 @@
 #include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
 #include <seqan3/core/simd/simd.hpp>
-#include <seqan3/core/type_traits/lazy.hpp>
 #include <seqan3/range/views/chunk.hpp>
 #include <seqan3/range/views/zip.hpp>
 #include <seqan3/utility/type_traits/function_traits.hpp>
+#include <seqan3/utility/type_traits/lazy_conditional.hpp>
 
 namespace seqan3::detail
 {
