+++ include/seqan3/range/views/minimiser.hpp
@@ -16,9 +16,9 @@
 #include <deque>
 
 #include <seqan3/core/detail/empty_type.hpp>
-#include <seqan3/core/type_traits/lazy.hpp>
 #include <seqan3/range/concept.hpp>
 #include <seqan3/range/views/detail.hpp>
+#include <seqan3/utility/type_traits/lazy_conditional.hpp>
 
 namespace seqan3::detail
 {
