+++ include/seqan3/utility/type_traits/lazy_conditional.hpp
@@ -156,44 +156,6 @@ template <bool decision, typename on_true_t, typename on_false_t>
 //!\endcond
 using lazy_conditional_t = instantiate_t<std::conditional_t<decision, on_true_t, on_false_t>>;
 
-/*!\brief An unary type trait that tests whether a template class can be declared with the given template type
- *        parameters.
- * \implements seqan3::unary_type_trait
- * \tparam query_t The type of the template class to test.
- * \tparam args_t  The template parameter pack to instantiate the template class with.
- *
- * \details
- *
- * Note, this unary type trait can be used in a seqan3::detail::lazy_conditional expression to check if instantiating
- * a template class with specific template arguments would result in a valid template declaration. Thus, the template
- * parameters of the checked class must be constrained accordingly. It is, however, not possible to test if the
- * the resulting type is incomplete or not, such that it can not be tested if an instance of the class template with
- * the given template arguments can be actually created.
- *
- * ### Example
- *
- * \include test/snippet/core/type_traits/is_class_template_declarable_with.cpp
- */
-template <template <typename ...> typename query_t, typename ...args_t>
-struct is_class_template_declarable_with :
-//!\cond
-    public std::false_type
-//!\endcond
-{};
-
-//!\cond
-template <template <typename ...> typename query_t, typename ...args_t>
-    requires requires { typename std::type_identity<query_t<args_t...>>::type; }
-struct is_class_template_declarable_with<query_t, args_t...> : public std::true_type
-{};
-//!\endcond
-
-/*!\brief Helper variable template for seqan3::detail::is_class_template_declarable_with.
- * \tparam query_t The type of the template class to test.
- * \tparam args_t  The template parameter pack to instantiate the template class with.
- * \relates seqan3::detail::is_class_template_declarable_with
- */
-template <template <typename ...> typename query_t, typename ...args_t>
-inline constexpr bool is_class_template_declarable_with_v = is_class_template_declarable_with<query_t, args_t...>::value;
 //!\}
+
 } // namespace seqan3::detail
